Removed metadata attributes "othertracks_"
Removed attributes that were conceptualized in the initial implementation of multiple track. We only need to keep "referencetrack_included" as the only one needed. We decided it is cleaner to maintain an eTUFF file that has a 1-to-1 relationship between the logged sensor data and a set of waypoints (track)