Added trimming by dates before writing eTUFF
If trim_data flag is set to TRUE, a global trimming of data to within start & end dates (specified by metadata) is done before writing to eTUFF file. Before this edit, various places in the code tried to do the same. But it is probably the best to do this "global" trimming right before writing to file.